+++ src/marsyas/AimLocalMax.cpp
@@ -118,8 +118,8 @@ AimLocalMax::myUpdate(MarControlPtr sender)
 
 bool
 AimLocalMax::InitializeInternal() {
-  strobe_timeout_samples_ = floor(ctrl_timeout_ms_->to<mrs_real>() * ctrl_israte_->to<mrs_real>() / 1000.0);
-  strobe_decay_samples_ = floor(ctrl_decay_time_ms_->to<mrs_real>() * ctrl_israte_->to<mrs_real>() / 1000.0);
+  strobe_timeout_samples_ = (int) floor(ctrl_timeout_ms_->to<mrs_real>() * ctrl_israte_->to<mrs_real>() / 1000.0);
+  strobe_decay_samples_ = (int) floor(ctrl_decay_time_ms_->to<mrs_real>() * ctrl_israte_->to<mrs_real>() / 1000.0);
   return true;
   
 }
